Split DispatchingGroup into own file
