+++ onTopPoke/Models/Dispatching.swift
@@ -9,14 +9,3 @@ extension DispatchQueue: Dispatching {
         async(execute: work)
     }
 }
-
-protocol DispatchingGroup {
-    func notify(queue: DispatchQueue,
-                _ work: @escaping () -> Void)
-}
-
-extension DispatchGroup: DispatchingGroup {
-    func notify(queue: DispatchQueue, _ work: @escaping () -> Void) {
-        self.notify(queue: queue, execute: work)
-    }
-}
